+++ core/impl/src/main/java/com/blazebit/persistence/impl/CriteriaBuilderConfigurationImpl.java
@@ -513,8 +513,10 @@ import com.blazebit.persistence.impl.function.window.percentrank.PercentRankFunc
 import com.blazebit.persistence.impl.function.window.rank.RankFunction;
 import com.blazebit.persistence.impl.function.window.row.RowNumberFunction;
 import com.blazebit.persistence.impl.function.window.sum.SumFunction;
+import com.blazebit.persistence.impl.util.CriteriaBuilderConfigurationContributorComparator;
 import com.blazebit.persistence.parser.expression.ConcurrentHashMapExpressionCache;
 import com.blazebit.persistence.spi.CriteriaBuilderConfiguration;
+import com.blazebit.persistence.spi.CriteriaBuilderConfigurationContributor;
 import com.blazebit.persistence.spi.DbmsDialect;
 import com.blazebit.persistence.spi.EntityManagerFactoryIntegrator;
 import com.blazebit.persistence.spi.ExtendedQuerySupport;
@@ -533,6 +535,7 @@ import java.sql.Time;
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.Collections;
 import java.util.GregorianCalendar;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -567,6 +570,7 @@ public class CriteriaBuilderConfigurationImpl implements CriteriaBuilderConfigur
         loadEntityManagerIntegrator();
         loadDbmsDialects();
         loadFunctions();
+        loadExtensions();
     }
 
     // NOTE: When adding a function here, you might want to also add it in AbstractCoreTest so it is recognized
@@ -1873,6 +1877,20 @@ public class CriteriaBuilderConfigurationImpl implements CriteriaBuilderConfigur
         }
     }
 
+    private void loadExtensions() {
+        List<CriteriaBuilderConfigurationContributor> contributors = new ArrayList<>();
+
+        for (CriteriaBuilderConfigurationContributor contributor : ServiceLoader.load(CriteriaBuilderConfigurationContributor.class)) {
+            contributors.add(contributor);
+        }
+
+        Collections.sort(contributors, new CriteriaBuilderConfigurationContributorComparator());
+
+        for (CriteriaBuilderConfigurationContributor contributor : contributors) {
+            contributor.contribute(this);
+        }
+    }
+
     @Override
     public CriteriaBuilderConfiguration withPackageOpener(PackageOpener packageOpener) {
         this.packageOpener = packageOpener;
@@ -1999,4 +2017,5 @@ public class CriteriaBuilderConfigurationImpl implements CriteriaBuilderConfigur
         properties.setProperty(propertyName, value);
         return this;
     }
+
 }