+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/update/natural/model/BookIsbnWithPkView.java
@@ -14,20 +14,21 @@
  * limitations under the License.
  */
 
-package com.blazebit.persistence.view.testsuite.basic.model;
+package com.blazebit.persistence.view.testsuite.update.natural.model;
 
+import com.blazebit.persistence.testsuite.entity.BookEntity;
 import com.blazebit.persistence.view.EntityView;
-import com.blazebit.persistence.view.testsuite.TypeUseAnnotation;
-import com.blazebit.persistence.view.testsuite.entity.NamedEntity;
+import com.blazebit.persistence.view.IdMapping;
+import com.blazebit.persistence.view.testsuite.basic.model.IdHolderView;
 
 /**
  *
  * @author Christian Beikov
- * @since 1.0.0
+ * @since 1.6.0
  */
-@EntityView(NamedEntity.class)
-public interface NamedView extends IdHolderView<Integer> {
-
-    @TypeUseAnnotation
-    public String getName();
+@EntityView(BookEntity.class)
+public interface BookIsbnWithPkView {
+    Long getId();
+    @IdMapping
+    String getIsbn();
 }