+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/update/natural/model/UpdatableBookReferencePkView.java
@@ -14,20 +14,32 @@
  * limitations under the License.
  */
 
-package com.blazebit.persistence.view.testsuite.basic.model;
+package com.blazebit.persistence.view.testsuite.update.natural.model;
 
+import com.blazebit.persistence.testsuite.entity.BookISBNReferenceEntity;
 import com.blazebit.persistence.view.EntityView;
-import com.blazebit.persistence.view.testsuite.TypeUseAnnotation;
-import com.blazebit.persistence.view.testsuite.entity.NamedEntity;
+import com.blazebit.persistence.view.IdMapping;
+import com.blazebit.persistence.view.UpdatableEntityView;
 
 /**
  *
  * @author Christian Beikov
- * @since 1.0.0
+ * @since 1.6.0
  */
-@EntityView(NamedEntity.class)
-public interface NamedView extends IdHolderView<Integer> {
+@UpdatableEntityView
+@EntityView(BookISBNReferenceEntity.class)
+public interface UpdatableBookReferencePkView {
+    
+    @IdMapping
+    public Long getId();
 
-    @TypeUseAnnotation
-    public String getName();
+    public Long getVersion();
+
+    public BookPkWithIsbnView getBook();
+
+    public void setBook(BookPkWithIsbnView book);
+
+    public BookIsbnWithPkView getBookNormal();
+
+    public void setBookNormal(BookIsbnWithPkView bookNormal);
 }