+++ CHANGELOG.md
@@ -8,10 +8,14 @@ Not yet released
 
 ### New features
 
-None yet
+* Introduction of the `CriteriaBuilderConfigurationContributors` SPI which allows third party libraries to extend the default `CriteriaBuilderConfiguration`.
+* Add support for Spring Boot/Data 2.5
 
 ### Bug fixes
-None yet
+
+* Fix support for Java 8 type temporal literals
+* Fix issues with persisting updatable non-PK subview mapping referring to PK-Id view
+* Fix issues with entity-view annotation processor and type use annotations
 
 ### Backwards-incompatible changes